app/requirements.txt
android_backup==0.2.0
appdirs==1.4.4
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
construct==2.10.70
croniter==6.0.0
cryptography==45.0.5
defusedxml==0.7.1
Flask==3.1.1
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
micloud==0.6
netifaces==0.11.0
pycparser==2.22
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
python-miio==0.5.12
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
tqdm==4.67.1
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
zeroconf==0.147.0
